refactor Questions component into QuestionList and Question components
